+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- Added `\Wizaplace\SDK\Company\CompanyService::updateFile`
+- Added `\Wizaplace\SDK\Company\CompanyService::deleteFile`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationFromUserId`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::addUserToGroup`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::removeUserToGroup`